refactor(modal): extract validation normalization helpers
- Move validation/normalization helpers from app.py to validation.py
- Functions moved: _to_isoformat, estimate_stage, parse_tags, normalize_tags,
  normalize_memory_row, normalize_tree_row, normalize_row, validate_visibility,
  validate_optional_string, validate_required_uuid, validate_optional_uuid
- app.py imports from validation.py, no behavior changes
- Pure extraction, preserves all function signatures and logic
